pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.3.0"
description = ""
authors = ["skip"]
readme = "README.md"
packages = [
{ include = "page_analyzer"},
]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
flask = "^3.0.3"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
